fine-tune: listen to deltas: - extract filterInsertedJobUris method KAS-3179 pub: v3: reports: basic
