+++ app.js
@@ -1,6 +1,7 @@
 import path from 'path';
 import mu from 'mu';
 import * as config from './config.js';
+import * as Delta from './lib/delta.js';
 import * as VirtuosoClient from './lib/VirtuosoClient.js';
 import * as RegulationType from './queries/reports/RegulationType.js';
 import bodyParser from 'body-parser';
@@ -23,26 +24,13 @@ mu.app.get('/', async function (req, res, next) {
   }
 });
 
-mu.app.post('/delta', bodyParser.json(), async function (req, res, next) {
+mu.app.post('/delta', bodyParser.json(), async function (req, res) {
   res.status(202).end();
 
   let deltas = req.body;
-  let newJobUris = [];
-  for (let delta of deltas) {
-    let inserts = delta.inserts;
-    for (let triple of inserts) {
-      let isJob =
-        triple.predicate.value === 'http://www.w3.org/1999/02/22-rdf-syntax-ns#type' &&
-        triple.object.value === 'http://mu.semte.ch/vocabularies/ext/publicatie/PublicationMetricsExportJob';
+  let newJobUris = Delta.filterInsertedJobUris(deltas);
 
-      if (isJob) {
-        let jobUri = triple.subject.value;
-        newJobUris.push(jobUri);
-      }
-    }
-  }
-
-  console.log("I received jobs with URI: ", newJobUris.join(' '))
+  console.log('I received jobs with URIs: ', newJobUris.join(' '));
 });
 
 mu.app.use(mu.errorHandler);
